# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "QuieroKetchup API",
        "description": "Documentación de endpoints públicos de QuieroKetchup.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar productos",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "subcategory_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "restaurant_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "in_promotion",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "available",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Ver un producto",
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Hamburguesas"
                    },
                    "slug": {
                        "type": "string",
                        "example": "hamburguesas"
                    },
                    "description": {
                        "type": "string",
                        "example": "",
                        "nullable": true
                    },
                    "meta": {
                        "type": "object",
                        "nullable": true
                    },
                    "subcategories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Subcategory"
                        }
                    }
                },
                "type": "object"
            },
...
            "ProductOptionItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_option_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Queso extra"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 5
                    },
                    "meta": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ProductOption": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Extras"
                    },
                    "type": {
                        "type": "string",
                        "example": "multiple",
                        "nullable": true
                    },
                    "required": {
                        "type": "boolean",
                        "example": false
                    },
                    "meta": {
                        "type": "object",
                        "nullable": true
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductOptionItem"
                        }
                    }
                },
                "type": "object"
            },
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "subcategory_id": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Hamburguesa Clásica"
                    },
                    "slug": {
                        "type": "string",
                        "example": "hamburguesa-clasica"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 49.9
                    },
                    "currency": {
                        "type": "string",
                        "example": "PYG",
                        "nullable": true
                    },
                    "available": {
                        "type": "boolean",
                        "example": true
                    },
                    "meta": {
                        "type": "object",
                        "nullable": true
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "subcategory": {
                        "$ref": "#/components/schemas/Subcategory"
                    },
                    "restaurants": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Restaurant"
                        }
                    },
                    "options": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductOption"
                        }
                    },
                    "promotions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Promotion"
                        }
                    }
                },
                "type": "object"
            },
            "Promotion": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "type": "string",
                        "example": "PROMO10",
                        "nullable": true
                    },
                    "type": {
                        "type": "string",
                        "example": "percentage",
                        "nullable": true
                    },
                    "value": {
                        "type": "number",
                        "format": "float",
                        "example": 10
                    },
                    "starts_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "ends_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "meta": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Restaurant": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sucursal Centro"
                    },
                    "slug": {
                        "type": "string",
                        "example": "sucursal-centro"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "nullable": true
                    },
                    "settings": {
                        "type": "object",
                        "nullable": true
                    },
                    "meta": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Subcategory": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Clásicas"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}